==> code/1 - Tokenize dữ liệu.ipynb <==
from google.colab import drive
drive.mount('/content/drive')

# Tải vncorenlp

!pip3 install vncorenlp
%mkdir -p vncorenlp/models/wordsegmenter
!wget https://raw.githubusercontent.com/vncorenlp/VnCoreNLP/master/VnCoreNLP-1.1.1.jar
!wget https://raw.githubusercontent.com/vncorenlp/VnCoreNLP/master/models/wordsegmenter/vi-vocab
!wget https://raw.githubusercontent.com/vncorenlp/VnCoreNLP/master/models/wordsegmenter/wordsegmenter.rdr
%mv VnCoreNLP-1.1.1.jar vncorenlp/
%mv vi-vocab vncorenlp/models/wordsegmenter/
%mv wordsegmenter.rdr vncorenlp/models/wordsegmenter/

import os
import csv
from tqdm import tqdm
from transformers import AutoTokenizer
from vncorenlp import VnCoreNLP

# Tokenize và trích xuất thực thể

## Hàm tokenize và trích xuất thực thể

def get_data(tsv_file_path, output_entities_extraction):
    rdrsegmenter = VnCoreNLP("vncorenlp/VnCoreNLP-1.1.1.jar", annotators="wseg,pos,ner", max_heap_size='-Xmx500m')

    X = []
    y = []

    with open(tsv_file_path, 'r', encoding='utf-8') as f:
        tsv_reader = csv.reader(f, delimiter='\t')
        rows = []

        for row in tqdm(tsv_reader):
            if len(row) >= 1:
                content = row[0]
                label = row[1]

                if not content.strip():
                    print("Skipping empty line")
                    continue

                try:
                    content = content.encode('utf-8', errors='replace').decode('utf-8')
                    annotated_text = rdrsegmenter.annotate(content)

                    entities = []
                    tokenized_text = []
                    for sentence in annotated_text['sentences']:
                        for word in sentence:
                            tokenized_text.append(word['form'])
                            if word['nerLabel'] != 'O':  # Not a non-entity
                                entities.append(word['form'])

                    keywords = ','.join(entities)
                    tokenized_content = ' '.join(tokenized_text)

                    X.append(tokenized_content)
                    y.append(label)

                    rows.append({'content': content, 'keywords': keywords})
                except Exception as e:
                    print(f"Error processing text: {content}")
                    print(f"Error: {e}")
                    continue

    with open(output_entities_extraction, 'w', encoding='utf-8', newline='') as f:
      writer = csv.DictWriter(f, fieldnames=['content', 'keywords'], delimiter='\t')
      writer.writeheader()
      writer.writerows(rows)


train_path = '/content/drive/MyDrive/Data Mining Field-oriented/data/raw/articles_training.tsv'
output_entities_extraction = '/content/drive/MyDrive/Data Mining Field-oriented/output/output_ner.tsv'
X_data, y_data = get_data(train_path)

## Kiểm chứng kết quả

print(X_data[8892])
print(y_data[8892])

## Xuất output

from google.colab import files

output_path = '/content/drive/MyDrive/Data Mining Field-oriented/consolidated/tokenized_articles_training.tsv'
os.makedirs(os.path.dirname(output_path), exist_ok=True)

with open(output_path, 'w', encoding='utf-8', newline='') as out_file:
    writer = csv.writer(out_file, delimiter='\t')
    for x, y in zip(X_data, y_data):
        writer.writerow([x, y])

files.download(output_path)
