==> code/6 - LDA mức 2.ipynb <==
# Kết nối Drive

from google.colab import drive
drive.mount('/content/drive')

# Đọc dữ liệu

import os
import pandas as pd

%cd /content/drive/MyDrive/Data Mining Field-oriented/data/consolidated/topic

file_name = f"Pháp_luật.tsv"
try:
    df = pd.read_csv(file_name, sep="\t", encoding="utf-8")
except FileNotFoundError:
    print(f"File {file_name} not found.")

# Xây dựng chủ đề ẩn bằng LDA

from gensim import corpora
from gensim.models import LdaModel

!pip3 install pyLDAvis
import pyLDAvis
import pyLDAvis.gensim_models as gensimvis

## Loại bỏ nhiễu khỏi văn bản

import re

def preprocess(text):
    text = str(text)
    pattern_keep = r'(?<=[a-zA-Z0-9])(\.|,)(?=[a-zA-Z0-9])|(?<=[a-zA-Z])\.(?=\s|,|$)'

    text = re.sub(pattern_keep, r'\1', text)

    pattern_remove = r'[.,](?=\s|$|[^a-zA-Z0-9])'
    isbn_pattern = r'\b(?:\d{9}[\dXx]|\d{13})\b'
    ip_pattern = r'\b(?:\d{1,3}\.){3}\d{1,3}\b|\b(?:[a-fA-F0-9]{1,4}:){7}[a-fA-F0-9]{1,4}\b'

    cleaned_text = re.sub(pattern_remove, ' ', text)
    cleaned_text = re.sub(isbn_pattern, ' ', cleaned_text)
    cleaned_text = re.sub(ip_pattern, ' ', cleaned_text)

    cleaned_text = re.sub(r'\s{2,}', ' ', cleaned_text)
    cleaned_text = re.sub(r'[/-]', ' ', cleaned_text)
    cleaned_text = re.sub(r'\b\d+\b', '', cleaned_text)
    cleaned_text = re.sub(r'\s{2,}', ' ', cleaned_text)

    return cleaned_text.strip().split()

## Tạo từ điển và ma trận Bag-of-Words

df['content'] = df['content'].astype(str)
documents = df['content'].apply(lambda x: x.split()).tolist()

dictionary = corpora.Dictionary(documents)
print(f"Số lượng từ trong từ điển: {len(dictionary)}")

corpus = [dictionary.doc2bow(doc) for doc in documents]
print(f"Ví dụ Bag-of-Words cho tài liệu đầu tiên: {corpus[0]}")

## Mô hình LDA

num_topics = 3

lda_model = LdaModel(corpus, num_topics=num_topics, id2word=dictionary, passes=10)

print("\nChủ đề LDA:")
for idx, topic in lda_model.print_topics(-1):
    print(f"Chủ đề {idx}:\n{topic}\n")

## Đánh giá

from gensim.models.coherencemodel import CoherenceModel

def calculate_perplexity(lda_model, corpus):
    return lda_model.log_perplexity(corpus)

def calculate_coherence(lda_model, texts, dictionary, coherence_type='c_v'):
    coherence_model = CoherenceModel(
        model=lda_model,
        texts=texts,
        dictionary=dictionary,
        coherence=coherence_type
    )
    return coherence_model.get_coherence()

results = []

num_topic = 3
lda_model = LdaModel(corpus, num_topics=num_topic, id2word=dictionary, passes=5)

# perplexity_score = calculate_perplexity(lda_model, corpus)

# coherence_score = calculate_coherence(lda_model, documents, dictionary, coherence_type='c_v')

results.append({
    "num_topics": num_topic,
    # "perplexity": perplexity_score,
    # "coherence_score": coherence_score
})

## Thống kê chủ đề

### Thống kê theo số lượng

from collections import Counter

def get_stat(lda_model, corpus):
    stat = []
    for doc_bow in corpus:
        topic_probs = lda_model.get_document_topics(doc_bow)
        stat_count = max(topic_probs, key=lambda x: x[1])[0]
        stat.append(stat_count)
    return stat

stat = get_stat(lda_model, corpus)

topic_counts = Counter(stat)

print("\nSố lượng tài liệu thuộc mỗi chủ đề:")
for topic, count in topic_counts.items():
    print(f"Chủ đề {topic}: {count} tài liệu")

print(f"\nTổng số tài liệu: {sum(topic_counts.values())}")

### Lấy ra 200 từ nổi bật

%cd /content/drive/MyDrive/Data Mining Field-oriented/stats/topic/subtopic

def export_topics_to_csv(lda_model, num_topics, dictionary, output_file):
    topics = {}
    for topic_id in range(num_topics):
        topic_words = lda_model.show_topic(topic_id, topn=200)
        topics[f"Chủ đề {topic_id}"] = [word for word, weight in topic_words]

    topic_df = pd.DataFrame.from_dict(topics, orient="index").transpose()

    topic_df.to_csv(output_file, index=False, encoding="utf-8-sig")

output_file = "topic_words_Pháp_luật.csv"
export_topics_to_csv(lda_model, num_topics, dictionary, output_file)

### Thống kê tài liệu theo các chủ đề ẩn

def get_documents_topic_distribution_with_content(lda_model, corpus, num_topics, content_list, top_n=5):
    data = []
    for doc_index, doc_bow in enumerate(corpus):
        topic_probs = lda_model.get_document_topics(doc_bow, minimum_probability=0)
        topic_probs_sorted = sorted(topic_probs, key=lambda x: x[1], reverse=True)[:top_n]

        row = {"Tài liệu": f"Tài liệu {doc_index + 1}", "Nội dung": content_list[doc_index]}
        for i, (topic_id, prob) in enumerate(topic_probs_sorted):
            row[f"Chủ đề {i + 1}"] = f"Chủ đề {topic_id} ({prob:.2%})"
        data.append(row)
    return pd.DataFrame(data)

content_list = df['content'].tolist()

top_n_topics = 5
documents_topic_df_with_content = get_documents_topic_distribution_with_content(
    lda_model, corpus, num_topics, content_list, top_n=top_n_topics
)

documents_with_content_output_file = "stats_for_Pháp_luật.csv"
documents_topic_df_with_content.to_csv(documents_with_content_output_file, index=False, encoding="utf-8-sig")
