==> code/9 - Gộp kết quả.ipynb <==
from google.colab import drive
drive.mount('/content/drive')

import pandas as pd
import glob
import os

folder_path = "/content/drive/MyDrive/Data Mining Field-oriented/output"

csv_files = glob.glob(os.path.join(folder_path, "*.csv"))

df_list = []

for file in csv_files:
    df = pd.read_csv(file)
    df_list.append(df)

combined_df = pd.concat(df_list, ignore_index=True)

print(combined_df.shape)

import pandas as pd
import glob
import os

df_list = []

for file in glob.glob("*.csv"):
    df = pd.read_csv(file)
    df_list.append(df)
    file_name = os.path.basename(file)
    print(f"Tệp: {file_name}, Kích thước: {df.shape}")

combined_df_sorted = combined_df.sort_values(by='idx', ascending=True)
combined_df_sorted['keywords'] = combined_df_sorted['keywords'].str.replace(', ', ',', regex=False)

print(combined_df_sorted)

file_path_ner = "/content/drive/MyDrive/Data Mining Field-oriented/output/output_ner.tsv"
df_ner = pd.read_csv(file_path_ner, encoding='utf-8-sig')

df_ner = df_ner[df_ner['idx'] >= 0]
print(df_ner)

import pandas as pd

idx_values = list(range(43903))

df = pd.DataFrame({
    'idx': idx_values,
    'keywords': [''] * len(idx_values)
})

print(df.info())

print(df)

import pandas as pd

combined_keywords = pd.concat([df_ner[['idx', 'keywords']], combined_df[['idx', 'keywords']]])
combined_keywords = combined_keywords.drop_duplicates(subset='idx', keep='last')
combined_keywords_sorted = combined_keywords.sort_values(by='idx', ascending=True)

print(combined_keywords.head())

print(combined_keywords)

import pandas as pd

combined_keywords_grouped = combined_keywords.groupby('idx', as_index=False).agg({
    'keywords': lambda x: ','.join(filter(None, map(str, x)))
})
combined_keywords_grouped['keywords'] = combined_keywords_grouped['keywords'].fillna('').str.replace('_', ' ', regex=False)
combined_keywords_grouped['keywords'] = combined_keywords_grouped['keywords'].fillna('').str.lower()
print(combined_keywords_grouped)

def remove_duplicate_keywords(keywords):
    if pd.isna(keywords):
        return keywords

    keywords_list = keywords.split(',')
    unique_keywords = list(dict.fromkeys(keywords_list))
    return ','.join(unique_keywords)

combined_keywords_grouped['keywords'] = combined_keywords_grouped['keywords'].apply(remove_duplicate_keywords)

print(combined_keywords_grouped)

merged_df = pd.merge(df, combined_keywords_grouped, on='idx', how='left')
merged_df = merged_df.T.drop_duplicates().T

print(merged_df)

import numpy as np

merged_df['keywords'] = merged_df['keywords_x'].fillna('') + ', ' + merged_df['keywords_y'].fillna('')
merged_df['keywords'] = merged_df['keywords'].str.strip(', ')
merged_df['keywords'] = merged_df['keywords'].replace('', np.nan)

merged_df.drop(columns=['keywords_x', 'keywords_y'], inplace=True)

print(merged_df)

df_test = pd.read_csv('/content/drive/MyDrive/Data Mining Field-oriented/data/raw/articles_testing.tsv', sep='\t')

merged_df.rename(columns={'keywords': 'tags'}, inplace=True)

combined_df = pd.concat([df_test, merged_df[['tags']]], axis=1)

print(combined_df)

combined_df.to_csv('nhom3_sol1.tsv', sep='\t', index=False, encoding='utf-8-sig')

from google.colab import files
files.download('nhom3_sol1.tsv')
