==> code/10 - SVM Binary Classification.ipynb <==
# Kết nối Drive

# Đọc dữ liệu

import pandas as pd

file_name = "/kaggle/input/hahaha/Thoi_trang_phong_cach.csv"
df = pd.read_csv(file_name, encoding="utf-8-sig")

df = df.drop(df.columns[3], axis=1)
print(df.head(3))

# Hàm xóa stop words, số, loại bỏ từ đơn

with open('/kaggle/input/hahaha/StopWord.txt', 'r', encoding='utf-8') as f:
  vietnamese_stopwords = [word.strip() for word in f.readlines() if word.strip()]

import re

def preprocess_text(text, stopwords):
    text = text.lower()
    cleaned_text = re.sub(r' [^a-zA-Z0-9\sÀ-ỹ|] ', ' ', text)
    cleaned_text = re.sub(r'[^a-zA-Z0-9\sÀ-ỹ|] ', ' ', cleaned_text)

    cleaned_text = re.sub(r'\|([^|]+)\|', r'\1', cleaned_text)

    cleaned_text = re.sub(r'\s{2,}', ' ', cleaned_text).strip()
    cleaned_text = re.sub(r'\b\d+\b', '', cleaned_text)
    cleaned_text = re.sub(r'[^a-zA-Z0-9\sÀ-ỹ|] ', ' ', cleaned_text)
    cleaned_text = re.sub(r'[,]', ' ', cleaned_text)

    words = cleaned_text.split()
    processed_words = [word for word in words if word not in stopwords]

    return ' '.join(processed_words)

preprocess_text("4 tp.H.C.M phát_động, tuy_nhiên thi Đại_sứ văn_hoá đọc 2024 thu_hút 1,6 triệu dự thi học_sinh, sinh_viên 9,2 ngàn cơ_sở giáo_dục nước tham_gia. 517 giải cao vòng sơ_khảo 60 tỉnh thành gửi Ban tổ_chức. Kết_quả, Ban tổ_chức trao 4 danh_hiệu Đại_sứ văn_hoá đọc tiêu_biểu 136 giải nhất, nhì, ba, khuyến_khích giải chuyên_đề thí_sinh tham_gia. đó, Đồng_Nai đoạt 2 giải khuyến_khích Cù_Đặng_Bảo_Trân, lớp 84, Trường TH-THCS-THPT Song_ngữ Á_Châu Nguyễn_Thái_Bình, lớp 12C1, Trường trung_học_phổ_thông Lê_Quý_Đôn Tân_Mai, thành_phố Biên_Hoà", vietnamese_stopwords)

# Chuẩn bị dữ liệu để nạp vào mô hình SVM

df['Nội dung'] = df['Nội dung'].apply(lambda x: preprocess_text(x, vietnamese_stopwords))
print(df['Nội dung'])

!pip3 install pyvi

import pandas as pd
import numpy as np
from pyvi import ViTokenizer, ViPosTagger
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.model_selection import train_test_split
from sklearn.preprocessing import OneHotEncoder
from sklearn.svm import LinearSVC
from sklearn.metrics import classification_report

def preprocess_and_extract_pos(tokens):
    _, pos_tags = ViPosTagger.postagging(" ".join(tokens))
    return pos_tags

pos_tag = df['Nội dung'].apply(preprocess_and_extract_pos)

import json

df['POS_Tags'] = pos_tag.apply(lambda x: x.split(',') if isinstance(x, str) else x)
df['POS_Tags'] = df['POS_Tags'].apply(lambda x: json.dumps(x) if isinstance(x, list) else x)

print(df.head())

# output_file_path = '/content/processed_dataframe.csv'
# df.to_csv(output_file_path, index=False, encoding='utf-8-sig')

# output_file_path

df['Tags'] = df['Tags'].str.lower()
df['Tags_split'] = df['Tags'].apply(lambda x: set(x.split(',')) if isinstance(x, str) else set())
df['Tags_split'] = df['Tags_split'].apply(
    lambda tags: {tag.replace(' ', '_') for tag in tags} if isinstance(tags, set) else tags
)

def prepare_training_data(row):
    tokens = row['Nội dung'].split()
    labels = [1 if token in row['Tags_split'] else 0 for token in tokens]
    return labels

df['Labels'] = df.apply(prepare_training_data, axis=1)

print(df.head())
tmp_df = df

from tqdm import tqdm

new_data = []

for index, row in tqdm(df.iterrows(), total=df.shape[0], desc="Processing rows"):
    tokens = row['Nội dung'].split()
    pos_tags = row['POS_Tags']
    for word, pos in zip(tokens, pos_tags):
        label = 1 if word in row['Tags_split'] else 0
        new_entry = {'idx': index, 'Từ': word, 'POS Tag': pos, 'Labels': label}
        new_data.append(new_entry)

new_df = pd.DataFrame(new_data)

print(new_df.head())

from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.model_selection import train_test_split
from sklearn.svm import LinearSVC
from sklearn.metrics import classification_report

vectorizer = TfidfVectorizer()
X = vectorizer.fit_transform(df['Nội dung'])

import pandas as pd
from sklearn.feature_extraction.text import TfidfVectorizer

terms = vectorizer.get_feature_names_out()

dense_matrix = X.toarray()
tfidf_scores = dense_matrix.sum(axis=0)

tfidf_df = pd.DataFrame({'Term': terms, 'TF-IDF': tfidf_scores}).sort_values(by='TF-IDF', ascending=False)
tfidf_df

df = tmp_df
tmp_new_df = new_df

print(new_df.shape)

import pandas as pd
from sklearn.preprocessing import MinMaxScaler
from sklearn.preprocessing import LabelEncoder
from sklearn.preprocessing import OneHotEncoder

tfidf_df.rename(columns={'Term': 'Từ'}, inplace=True)

label_encoder = LabelEncoder()

merged_df = pd.merge(new_df, tfidf_df, on='Từ', how='left')
merged_df.rename(columns={'TF-IDF': 'TF-IDF Value'}, inplace=True)
merged_df = merged_df.dropna(subset=['TF-IDF Value'])

scaler = MinMaxScaler()
merged_df['TF-IDF Value'] = scaler.fit_transform(merged_df[['TF-IDF Value']])

one_hot_encoder = OneHotEncoder(sparse=False)
encoded_pos = one_hot_encoder.fit_transform(merged_df[['POS Tag']])

pos_columns = one_hot_encoder.get_feature_names_out(['POS Tag'])  # Lấy tên các cột sau encoding
encoded_pos_df = pd.DataFrame(encoded_pos, columns=pos_columns, index=merged_df.index)

merged_df = pd.concat([merged_df, encoded_pos_df], axis=1)

merged_df = merged_df.drop(columns=['POS Tag'])

columns_to_keep = ['idx', 'Từ', 'TF-IDF Value', 'Labels', 'POS Tag_N', 'POS Tag_V']
merged_df = merged_df[columns_to_keep]

labels = merged_df.pop('Labels')
merged_df['Labels'] = labels

print(merged_df)

label_counts = merged_df['Labels'].value_counts(normalize=True) * 100

label_distribution = pd.DataFrame({
    'Label': label_counts.index,
    'Percentage': label_counts.values
})

print(label_distribution)

import pandas as pd
from sklearn.model_selection import train_test_split
from sklearn.preprocessing import StandardScaler
from sklearn.svm import SVC
from sklearn.metrics import accuracy_score

X = train_columns = merged_df.iloc[:, 2:-1]
y = merged_df['Labels'] 

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=42)

class_weights = {0: 1 - (86 / 100), 1: 1 - (14 / 100)}

svm_model = SVC(kernel='linear', class_weight=class_weights)
svm_model.fit(X_train, y_train)

y_pred = svm_model.predict(X_test)

accuracy = accuracy_score(y_test, y_pred)
print(f'Độ chính xác của mô hình: {accuracy:.2f}')

report = classification_report(y_test, y_pred, target_names=["Non-Keyword", "Keyword"])
print(report)
